requirements-upgrade.txt
# Autonomous Survival Upgrade Dependencies

# Blockchain & Web3
web3>=6.0.0
eth-account>=0.8.0
eth-utils>=2.3.0

# CrewAI & LLM Support
crewai>=0.28.0
crewai-tools>=0.2.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
langchain-community>=0.0.20

# LLM Providers
openai>=1.0.0
anthropic>=0.18.0
ibm-watsonx-ai>=0.2.0  # WatsonX.ai SDK
ollama>=0.1.0  # Ollama Python client (optional)

# Environment & Config
python-dotenv>=1.0.0
pydantic>=2.5.0

# HTTP & API
requests>=2.28.0

# CLI Tools (for Akash integration)
# Note: Akash CLI should be installed separately via:
# curl -sSfL https://raw.githubusercontent.com/akash-network/node/master/install.sh | sh

# Payment Processing (optional - for production)
stripe>=5.0.0  # If using Stripe for fiat → crypto bridge
